+++ Smart_AI_backend/utils/errors/ForbiddenError.js
@@ -1,8 +1,8 @@
 const AppError = require('./AppError');
 
 class ForbiddenError extends AppError {
-  constructor(message = 'Bạn không có quyền truy cập chức năng này', code = 'FORBIDDEN') {
-    super(message, 403, code);
+  constructor(message = 'Bạn không có quyền truy cập chức năng này', code = 'FORBIDDEN', responseFormat = 'centralized') {
+    super(message, 403, code, undefined, responseFormat);
   }
 }