+++ Smart_AI_backend/utils/errors/UnauthorizedError.js
@@ -1,8 +1,8 @@
 const AppError = require('./AppError');
 
 class UnauthorizedError extends AppError {
-  constructor(message = 'Vui lòng đăng nhập để truy cập', code = 'UNAUTHORIZED') {
-    super(message, 401, code);
+  constructor(message = 'Vui lòng đăng nhập để truy cập', code = 'UNAUTHORIZED', responseFormat = 'centralized') {
+    super(message, 401, code, undefined, responseFormat);
   }
 }